+++ src/main/java/com/coin/exchange/bitcola/domain/BitcolaPair.java
@@ -14,7 +14,8 @@ public class BitcolaPair {
     //交易对名字
     private String pair;
     //数量精度
-    private BigDecimal amountScale;
+    private double amountScale;
     //价格精度
-    private BigDecimal priceScale;
+    private double priceScale;
+
 }
