+++ src/main/java/com/coin/facade/response/Pair.java
@@ -14,7 +14,7 @@ public class Pair {
     //交易对名字
     private String pair;
     //数量精度
-    private BigDecimal amountScale;
+    private double amountScale;
     //价格精度
-    private BigDecimal priceScale;
+    private double priceScale;
 }
